Added tests, edited Copy

+++ TM_CommonPy/__init__.py
@@ -3,6 +3,7 @@ import io
 import importlib
 import pip
 import xml.etree.ElementTree
+import shutil
 
 #Maybe use __file__ instead?
 def GetScriptRoot():
@@ -21,23 +22,28 @@ def GetXMLNamespaces(sXMLFile):
     ])
     return cNamespaces
 
+def Copy(src,root_dst_dir):
+    if os.path.isdir(src):
+        for src_dir, dirs, files in os.walk(src):
+            dst_dir = src_dir.replace(src, root_dst_dir, 1)
+            if not os.path.exists(dst_dir):
+                os.makedirs(dst_dir)
+            for file_ in files:
+                src_file = os.path.join(src_dir, file_)
+                dst_file = os.path.join(dst_dir, file_)
+                if os.path.exists(dst_file):
+                    os.remove(dst_file)
+                shutil.copy(src_file, dst_dir)
+    elif os.path.isfile(src):
+        dst_dir = src_dir.replace(src, root_dst_dir, 1)
+        shutil.copy(src, dst_dir)
+    else:
+        print("Copy\Error\src "+src+" is not a valid file or directory")
+
 #not tested
 def TryMakeDirs(sDir):
     if not os.path.exists(sDir):
         os.makedirs(sDir)
-
-#not tested
-def copyanything(root_src_dir,root_dst_dir):
-    for src_dir, dirs, files in os.walk(root_src_dir):
-        dst_dir = src_dir.replace(root_src_dir, root_dst_dir, 1)
-        if not os.path.exists(dst_dir):
-            os.makedirs(dst_dir)
-        for file_ in files:
-            src_file = os.path.join(src_dir, file_)
-            dst_file = os.path.join(dst_dir, file_)
-            if os.path.exists(dst_file):
-                os.remove(dst_file)
-            shutil.copy(src_file, dst_dir)
 
 #not tested
 def InstallAndImport(package):
